+++ main.lua
@@ -44,7 +44,8 @@ function love.load()
     -- "static" would be used for short sound effects (keeps it in memory)
     local music = love.audio.newSource("src/data/audio/chef_music.mp3", "stream")
     music:setLooping(true) -- Make it repeat forever
-    music:setVolume(0.5)   -- Set volume (0.0 to 1.0)
+    -- music:setVolume(0.5)   -- Set volume (0.0 to 1.0)
+    music:setVolume(0.0)   -- FIXME: Muted for now, uncomment out line above to enable music and delete this line
     music:play()